returned back empty lines, deleted unused grpc utils
