+++ hivemind/moe/client/beam_search.py
@@ -385,7 +385,7 @@ class MoEBeamSearcher:
             ),
             return_future,
         )
-        print(result)
+
         if return_future:
             return RemoteExpertWorker.spawn_experts_bulk_future(result, self.dht)
         return RemoteExpertWorker.spawn_experts_bulk(result, self.dht)
